fix compile errors

+++ src/libs/cpp/functions/index.cpp
@@ -11,7 +11,7 @@ void index(Context &context) noexcept
 		auto rhs = params[1];
 
 		// Find the index of the rhs in the lhs
-		for (int i = 0; i < lhs.size(); i++)
+		for (size_t i = 0; i < lhs.size(); i++)
 		{
 			if (lhs[i] == rhs)
 			{
@@ -26,7 +26,7 @@ void index(Context &context) noexcept
 		auto rhs = params[1].to_string();
 
 		// Find the index of the rhs in the lhs
-		int i = lhs.find(rhs);
+		size_t i = lhs.find(rhs);
 		if (i != std::string::npos)
 		{
 			index = i + 1;
